% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basic_scd}
\alias{basic_scd}
\title{Basic Single Case Design}
\format{
\subsection{\code{basic_scd}}{

A data frame with 21 rows and 3 columns:
\describe{
\item{phase}{Study phase}
\item{time}{Time of data collection}
\item{socbehavs}{Social behaviors score}
}
}
}
\source{
\url{https://ies.ed.gov/ncee/WWC/Docs/referenceresources/Final_WWC-HandbookVer5_0-0-508.pdf}
}
\usage{
basic_scd
}
\description{
This is the data for the Basic Single Case Design Example presented in
Figure 14 of the What Works Clearinghouse Procedures and Standards Handbook,
Version 5.0
}
\keyword{datasets}
