% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/napjack_helpers.R
\name{deal_phase}
\alias{deal_phase}
\title{Deal a Phase of Cards}
\usage{
deal_phase(.deck)
}
\arguments{
\item{.deck}{A deck of cards created using the \code{mmcards} package.}
}
\value{
A list with the following components:
\describe{
\item{cards_matrix}{A matrix of the dealt cards, with 1 row and 6 columns.}
\item{updeck}{The updated deck after dealing the cards.}
\item{dealt_cards}{A list of the individual dealt cards.}
}
}
\description{
This function deals a phase of cards from a given deck, consisting of 6 cards.
It returns a list containing the dealt cards as a matrix, the updated deck,
and the individual dealt cards.
}
\keyword{internal}
