% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{efficacy_of_CBT}
\alias{efficacy_of_CBT}
\title{Fictional Single Case Design Efficacy of CBT Example}
\format{
\subsection{\code{efficacy_of_CBT}}{

A data frame with 10 rows and 4 columns:
\describe{
\item{phase}{Study phase}
\item{time}{Time of data collection}
\item{Anxious}{Outcome measure}
\item{CATS_N}{Outcome measure}
}
}
}
\source{
\url{https://doi.org/10.4324/9780429273872}
}
\usage{
efficacy_of_CBT
}
\description{
This is the data set used for the clinical case example in Maric & van der Werf
(2020).
}
\references{
Maric, M., & van der Werff, V. (2020). Single-Case Experimental
Designs in Clinical Intervention Research. In R. van de Schoot & M. Miocevic
(Eds.), Small Sample Size Solutions: A Guide for Applied Researchers and
Practitioners (1st ed., pp. 10). Routledge.
https://doi.org/10.4324/9780429273872
}
\keyword{datasets}
