% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reversal_withdrawal}
\alias{reversal_withdrawal}
\title{Reversal/withdrawal Design Example}
\format{
\subsection{\code{reversal_withdrawal}}{

A data frame with 21 rows and 4 columns:
\describe{
\item{phase}{Study phase}
\item{time}{Time of data collection}
\item{extbehavs}{Externalizing behaviors score}
}
}
}
\source{
\url{https://ies.ed.gov/ncee/WWC/Docs/referenceresources/Final_WWC-HandbookVer5_0-0-508.pdf}
}
\usage{
reversal_withdrawal
}
\description{
This is the data for the Reversal/Withdrawal Design Example presented in
Figure 17 of the What Works Clearinghouse Procedures and Standards Handbook,
Version 5.0
}
\keyword{datasets}
